+++ src/spec/pages/Login.spec.tsx
@@ -3,10 +3,8 @@ import { MemoryRouter } from 'react-router';
 import renderer from 'react-test-renderer';
 import { act } from 'react-dom/test-utils';
 import { mount } from 'enzyme';
-import { AuthenticationDetails, CognitoUserPool } from 'amazon-cognito-identity-js';
 import Login from '@pages/Login';
 import Form from '@components/Form';
-import Input from '@components/Input';
 import Spinner from '@components/Spinner';
 
 describe('Login', () => {
@@ -18,7 +16,7 @@ describe('Login', () => {
   describe('render', () => {
     const testRenderer = renderer.create(
       <MemoryRouter>
-        <Login />
+        <Login setIsAuthorized={jest.fn()} />
       </MemoryRouter>
     );
 
@@ -32,7 +30,7 @@ describe('Login', () => {
     it('should display the spinner component after the submit button is clicked', () => {
       const wrapper = mount(
         <MemoryRouter>
-          <Login />
+          <Login setIsAuthorized={jest.fn()} />
         </MemoryRouter>
       );
       act(() => wrapper.find(Form).props().onSubmit());